add goto command, disable motion timeouts
A goto command gets a response from the motion controller immediately.
However, the next goto or track command will not be responded to
until the first goto finishes.  Thus (a) timeouts are inappropriate, and
(b) with no timeouts, issuing the second command blocks the event loop,
so it should be avoided if possible, e.g. by polling the position.