don't allow motion until mount has been zeroed
The handpad now has 'zero' and 'start/stop tracking' assigned
to the two mode buttons.  Allow handpad slewing but disable
remote goto, park, or read position until the mount has been
zeroed.  It can be zeroed at the handpad or remotely.

Also don't start initial tracking until requested (remotely or
via the handpad).  This is not because it is a danger so much
as it may be annoying if the mount tracks while trying to find
the zero.