standardize variable names, define origin
To be clear that motiond only handles raw telescope position,
not anything to do with the sky, name position coordinates
to (t,d), in arcseconds.

Motion controller cooordinates are named (x,y), in full steps.

Define the origin to be motion (0,0) and position (0,90*60*60),
the latter to roughly align with (LHA,DEC), e.g. pointing at
the polar axis and local meridian, telescope on the west side
of the pier, DEC shaft horizontal and out of the way of the
polar alignment scope.

Change the [ra] and [dec] stanzas in the config file to [t_axis]
and [d_axis] and the various variables that refer to the axes as
t and d instead of ra and dec.

Configure the (t,d) offset from (x,y) as 'offset' in the
config file.