enable monitoring of telescope position in gnuplot
'gem plot' plots the absolute telescope position in gnuplot, updating
it in real time as the telescope moves.

The graph is a polar plot using t as the angle and d as the magnitude.
gnuplot is launched by the utility.  When the user terminates the utility
e.g. with SIGINT, the gnuplot instance is terminated.

Position is polled once every 10 seconds while idle or tracking, once
every 1 second while moving (e.g. executing a GOTO or PARK).