regularly publish position information
Add a new zeromq pub/sub socket.  Add a timer to motiond, and
when it fires, publish the telescope position.  When idle/tracking
do this every 5s, but increase frequency when a move is in progress.

Modify 'gem plot' so that it subscribes to these updates and
passively waits for position updates, pushing them to the gnplot window
at whatever frequency they are offered.