add 'f' axis
Add axis to config data structure so we can have an [f_axis] stanza
in the config file.
- define 'steps' to be steps per meter of focus travel
  (compare steps/rev for t, d axes)
- define low_limit, high_limit to be in microns
  (compare arcsec to t, d axes)

Arrange for motiond to read config stanza and initialize f axis
according to configuration.

Add conversion functions for microns to steps and steps to microns.
We will use microns in the wire protocol encoding.