daemon: add --west option
The RA motor runs with a positive tracking velocity when
the telescope is on the east side of the pier, observing
west of the meridian.  It runs with a negative velocity
when tracking with the telescope on the west side of the
pier, observing east of the meridian.

Make the default be "observing east", and add a --west
option to reverse the polarity of the RA axis for
"merdian flip".

Drop the sign from the 'steps' attribute on the t axis
config file.