bbox: scale "encoder resolution" to 16K
Limit "encoder resolution" to 16K and scale positions
accordingly if the resolution exceeds that.
This is just in case returning more than 5 digits of
precision is a problem for some clients, or in case
a 16-bit short is being used to store values.

The decreased resolution won't matter much for syncing
the position of the telescope with a star map, which
is the primary goal of the bbox interface.