config: set "guide" rate
In preparation for having more configured slew rates,
rename the per-axis "slow" rate to "guide" and lower it
to a more reasonable value.

N.B. the lx200 protocol guide notes that the guide
rate should not be more than the sidereal rate, but
we can't clock DEC at a rate that low with the current
motion controller config, so go with a value slightly
higher than sidereal.  If we need to sort this out we
can change the DEC axis configuration.