hpad: use common slew parameters
Modify hpad module ot use common slew bits, and also
split out control (M1 + M2) keys and velocity.

Refactor "slew event" handling code in daemon.c so that the
slew start/stop logic is in one place.  It should now be possible
to simultaneously slew both axes without terminating one axis
when the button for the other is released.  (While the hpad
wiring does not allow this, lx200 and guide do allow it).

Ensure that the sidereal tracking speed, if tracking is enabled,
is added to the RA slew speed, especially important when guiding
in RA.