motion: switch to reactive model
Switch to reactive programming model internally.

Drop code to detect whether motion controllers
have been previously initialized, and instead
unconditionally perform a software reset and
load configuration parameters.  RA tracking is
always disabled on startup.

Open serial port in non-blocking mode and perform
all reads through an event loop io watcher.

If a response needs to be read synchronously,
do it, with timeout, in a temporary reactor that
only has the io watcher and timer active.

Implement a reactor based poller to monitor
motion status during a goto, and add a callback
to the daemon, so it can re-enable tracking
when the goto completes.

lx200: translate :Q# to soft stop via a callback.