daemon: refuse goto if not within +/-90*
It's probably safe to presume that the telescope is zeroed
within 90* of its furthest reach on either axis.  Therefore,
refuse a goto that is outside of that range.

Running down a bug in the pointing code that occasionally
tries to take one axis on a very long trip the wrong way round.
This defensive posture will be OK even after the bug is found.